+++ src/main/java/com/br/M1S2/miniprojeto02modulo01/services/MatricularServiceImpl.java
@@ -1,39 +1,58 @@
 package com.br.M1S2.miniprojeto02modulo01.services;
 import com.br.M1S2.miniprojeto02modulo01.entities.AlunoEntity;
+import com.br.M1S2.miniprojeto02modulo01.entities.DTOrequest.DTOrequest;
 import com.br.M1S2.miniprojeto02modulo01.entities.DisciplinaEntity;
 import com.br.M1S2.miniprojeto02modulo01.entities.DisciplinaMatriculaEntiy;
+import com.br.M1S2.miniprojeto02modulo01.exception.NotFoundException;
+import com.br.M1S2.miniprojeto02modulo01.repository.AlunoRepository;
 import com.br.M1S2.miniprojeto02modulo01.repository.DisciplinaMatriculaRepository;
+import com.br.M1S2.miniprojeto02modulo01.repository.DisciplinaRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class MatricularServiceImpl implements MatricularService {
 
     private final DisciplinaMatriculaRepository repository;
     private final AlunoService alunoService;
+    private final AlunoRepository alunoRepository;
+    private final DisciplinaRepository disciplinaRepository;
     private final DisciplinaService disciplinaService;
+    private static final Logger logger = LoggerFactory.getLogger(MatricularServiceImpl.class);
 
-    public MatricularServiceImpl(DisciplinaMatriculaRepository repository, AlunoService alunoService, DisciplinaService disciplinaService, NotasService notasService) {
+    public MatricularServiceImpl(DisciplinaMatriculaRepository repository, AlunoService alunoService, DisciplinaService disciplinaService, NotasService notasService, AlunoRepository alunoRepository, DisciplinaRepository disciplinaRepository) {
         this.repository = repository;
         this.alunoService = alunoService;
         this.disciplinaService = disciplinaService;
+        this.alunoRepository = alunoRepository;
+        this.disciplinaRepository = disciplinaRepository;
     }
 
     @Override
-    public DisciplinaMatriculaEntiy matricular(DisciplinaMatriculaEntiy matricula) {
-        //ID da matricula gerado automaticamente
+    public DisciplinaMatriculaEntiy matricular(DTOrequest entity) {
+        var aluno = alunoService.getById(entity.getAluno_id());
+        var disciplina = disciplinaService.getById(entity.getDisciplina_id());
+
+        if (!alunoRepository.existsById(aluno.getId())) {
+            logger.warn("Aluno não encontrado com ID: {}", aluno.getId());
+            throw new NotFoundException("Aluno não encontrado com ID: " + aluno.getId());
+        }
+
+        if (!disciplinaRepository.existsById(disciplina.getId())) {
+            logger.warn("Disciplina não cadastrada");
+            throw new NotFoundException("Disciplina não cadastrada");
+        }
+
+        DisciplinaMatriculaEntiy matricula = new DisciplinaMatriculaEntiy();
         matricula.setId(null);
-
-        //Seta para a matrícula o aluno indicado pelo ID
-        AlunoEntity aluno = alunoService.getById((matricula.getAluno().getId()));
         matricula.setAluno(aluno);
-
-        //Seta para a matrícula a disciplina indicada pelo ID
-        DisciplinaEntity disciplina = disciplinaService.getById((matricula.getDisciplina().getId()));
         matricula.setDisciplina(disciplina);
+        logger.info("Matricula concluida.");
 
-        //Salva em DisciplinaMatriculaEntity
         return repository.save(matricula);
     }
 
@@ -42,38 +61,68 @@ public class MatricularServiceImpl implements MatricularService {
     public void dell(Long id) throws Exception {
         DisciplinaMatriculaEntiy matricula = getById(id);
 
-        //Busca a lista de notas da matrícula a ser excluída
-        var notas = matricula.getNotas();
+        if (!repository.existsById(matricula.getId())) {
+            logger.warn("Matricula não encontrada");
+            throw new NotFoundException("Matricula não encontrada");
+        }
 
-        //Filtro - se há notas cadastradas, lança o erro
+        var notas = matricula.getNotas();
         if (!notas.isEmpty()){
             throw new Exception("Impossível excluir, há notas cadastradas!");
         }
+        logger.warn("Matricula excluida");
 
-        //Estando vazio, exclui
         repository.delete(matricula);
     }
-
 
     //Busca matricula pelo ID informado
     @Override
     public DisciplinaMatriculaEntiy getById(Long id) {
+        if (!repository.existsById(id)) {
+            logger.warn("Matricula não encontrada com ID: {}", id);
+            throw new NotFoundException("Matricula não encontrada com ID: " + id);
+        }
+
         return repository.findById(id).get();
     }
 
     //Busca as matrículas do aluno de ID informado
     @Override
     public List<DisciplinaMatriculaEntiy> getMatriculasByAlunoId(Long id) {
+        if (!alunoRepository.existsById(id)) {
+            logger.warn("Aluno não encontrado com ID: {}", id);
+            throw new NotFoundException("Aluno não encontrado com ID: " + id);
+        }
+
         var aluno = alunoService.getById(id);
-        return null; //aluno.getDisciplinas();
+
+        List<DisciplinaMatriculaEntiy> list = repository.findAll().stream().filter(x -> x.getAluno()
+                .getId().equals(aluno.getId())).collect(Collectors.toList());
+        return list;
     }
 
     //Retornar todas as matrículas de um disciplina
     @Override
     public List<DisciplinaMatriculaEntiy> getTodasDisciplinas(Long id) {
         var matricula = repository.getById(id);
+
+        if (!repository.existsById(matricula.getId())) {
+            logger.warn("Não encontrada a matricula de ID: {}", id);
+            throw new NotFoundException("Não encontrada a matricula de ID: " + id);
+        }
+
         var disciplina = matricula.getDisciplina();
-        return null; //disciplina.getMatriculas();
+        if (!repository.existsById(disciplina.getId())) {
+            logger.warn("Não cadastrada a disciplina de ID: {}", id);
+            throw new NotFoundException("Não cadastrada a disciplina de ID: " + id);
+        }
+
+        List<DisciplinaMatriculaEntiy> list = repository.findAll().stream().filter(x -> x.getDisciplina().getId()
+                .equals(matricula.getDisciplina().getId())).collect(Collectors.toList());
+
+        logger.info("Retornando todas as matrículas da disciplina de ID: {}", id);
+        return list;
     }
+
 
 }
